update for PR comments
