+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/queue/clear/DefaultTableClearer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.atlasdb.sweep.queue;
+package com.palantir.atlasdb.sweep.queue.clear;
 
 import java.util.Collection;
 import java.util.Map;
@@ -26,33 +26,28 @@ import java.util.stream.Collectors;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.LoadingCache;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import com.palantir.atlasdb.keyvalue.api.Cell;
-import com.palantir.atlasdb.keyvalue.api.CheckAndSetException;
-import com.palantir.atlasdb.keyvalue.api.CheckAndSetRequest;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.RangeRequest;
-import com.palantir.atlasdb.keyvalue.api.RowResult;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.api.Value;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.SweepStrategy;
-import com.palantir.atlasdb.ptobject.EncodingUtils;
-import com.palantir.atlasdb.schema.generated.TableClearsTable.TableClearsNamedColumn;
-import com.palantir.atlasdb.schema.generated.TableClearsTable.TableClearsRow;
-import com.palantir.atlasdb.schema.generated.TableClearsTable.TableClearsRowResult;
-import com.palantir.atlasdb.schema.generated.TargetedSweepTableFactory;
+import com.palantir.atlasdb.sweep.queue.ImmutableTimestampSupplier;
+import com.palantir.atlasdb.sweep.queue.TargetedSweepFilter;
+import com.palantir.atlasdb.sweep.queue.WriteInfo;
 import com.palantir.atlasdb.table.description.TableMetadata;
-import com.palantir.common.streams.KeyedStream;
 
 public final class DefaultTableClearer implements TableClearer, TargetedSweepFilter {
-    private static final TableReference CLEARS = TargetedSweepTableFactory.of().getTableClearsTable(null).getTableRef();
     private final LoadingCache<TableReference, SweepStrategy> sweepStrategies;
     private final KeyValueService kvs;
     private final ImmutableTimestampSupplier immutableTimestampSupplier;
+    private final ConservativeSweepWatermarkStore watermarkStore;
 
-    public DefaultTableClearer(KeyValueService kvs, ImmutableTimestampSupplier immutableTimestampSupplier) {
+    @VisibleForTesting
+    DefaultTableClearer(
+            ConservativeSweepWatermarkStore watermarkStore,
+            KeyValueService kvs,
+            ImmutableTimestampSupplier immutableTimestampSupplier) {
+        this.watermarkStore = watermarkStore;
         sweepStrategies = Caffeine.newBuilder()
                 .maximumSize(10_000)
                 .build(table -> Optional.ofNullable(kvs.getMetadataForTable(table))
@@ -64,6 +59,10 @@ public final class DefaultTableClearer implements TableClearer, TargetedSweepFil
         this.immutableTimestampSupplier = immutableTimestampSupplier;
     }
 
+    public DefaultTableClearer(KeyValueService kvs, ImmutableTimestampSupplier immutableTimestampSupplier) {
+        this(new DefaultConservativeSweepWatermarkStore(kvs), kvs, immutableTimestampSupplier);
+    }
+
     @Override
     public Collection<WriteInfo> filter(Collection<WriteInfo> writeInfos) {
         Set<TableReference> tablesToCareAbout = getConservativeTables(
@@ -73,23 +72,11 @@ public final class DefaultTableClearer implements TableClearer, TargetedSweepFil
             return writeInfos;
         }
 
-        Map<TableReference, Long> truncationTimes = getWatermarks(tablesToCareAbout);
+        Map<TableReference, Long> truncationTimes = watermarkStore.getWatermarks(tablesToCareAbout);
         return writeInfos.stream()
                 .filter(write -> !truncationTimes.containsKey(write.tableRef())
                         || write.timestamp() > truncationTimes.get(write.tableRef()))
                 .collect(Collectors.toList());
-    }
-
-    @VisibleForTesting
-    Map<TableReference, Long> getWatermarks(Set<TableReference> tableReferences) {
-        Set<Cell> cells = tableReferences.stream().map(DefaultTableClearer::cell).collect(Collectors.toSet());
-        Map<Cell, Value> fetched = kvs.get(CLEARS, Maps.asMap(cells, ignored -> Long.MAX_VALUE));
-        return KeyedStream.stream(fetched)
-                .map((cell, value) -> RowResult.of(cell, value.getContents()))
-                .map(TableClearsRowResult::of)
-                .mapKeys((cell, rowResult) -> tableRef(rowResult.getRowName()))
-                .map(TableClearsRowResult::getLastClearedTimestamp)
-                .collectToMap();
     }
 
     private Set<TableReference> getConservativeTables(Set<TableReference> tables) {
@@ -98,15 +85,6 @@ public final class DefaultTableClearer implements TableClearer, TargetedSweepFil
                         .map(SweepStrategy.CONSERVATIVE::equals)
                         .orElse(false))
                 .collect(Collectors.toSet());
-    }
-
-    private static Cell cell(TableReference table) {
-        TableClearsRow row = TableClearsRow.of(table.getQualifiedName());
-        return Cell.create(row.persistToBytes(), TableClearsNamedColumn.LAST_CLEARED_TIMESTAMP.getShortName());
-    }
-
-    private static TableReference tableRef(TableClearsRow row) {
-        return TableReference.fromString(row.getTable());
     }
 
     @Override
@@ -138,42 +116,6 @@ public final class DefaultTableClearer implements TableClearer, TargetedSweepFil
 
         destructiveAction.accept(conservativeTables);
 
-        updateWatermarksForTables(immutableTimestamp, conservativeTables);
-    }
-
-    @VisibleForTesting
-    void updateWatermarksForTables(long newWatermark, Set<TableReference> conservativeTables) {
-        conservativeTables.forEach(table -> updateWatermarkForTable(newWatermark, table));
-    }
-
-    private void updateWatermarkForTable(long newWatermark, TableReference table) {
-        while (true) {
-            Cell cell = cell(table);
-            Optional<Value> currentValue = Optional.ofNullable(
-                    kvs.get(CLEARS, ImmutableMap.of(cell, Long.MAX_VALUE)).get(cell));
-            Optional<Long> currentWatermark = currentValue.map(value -> RowResult.of(cell, value.getContents()))
-                    .map(TableClearsRowResult::of)
-                    .map(TableClearsRowResult::getLastClearedTimestamp);
-
-            if (currentWatermark.isPresent() && currentWatermark.get() > newWatermark) {
-                return;
-            }
-
-            CheckAndSetRequest request = currentWatermark
-                    .map(watermark -> CheckAndSetRequest.singleCell(
-                            CLEARS,
-                            cell,
-                            EncodingUtils.encodeVarLong(watermark),
-                            EncodingUtils.encodeVarLong(newWatermark)))
-                    .orElseGet(() -> CheckAndSetRequest.newCell(
-                            CLEARS, cell, EncodingUtils.encodeVarLong(newWatermark)));
-
-            try {
-                kvs.checkAndSet(request);
-                return;
-            } catch (CheckAndSetException e) {
-                // do nothing, we spin
-            }
-        }
+        watermarkStore.updateWatermarks(immutableTimestamp, conservativeTables);
     }
 }
